more version-fake data. fully implemented compare feature

+++ src/main/java/de/failender/dsaonline/rest/helden/Unterscheidbar.java
@@ -1,8 +1,6 @@
 package de.failender.dsaonline.rest.helden;
 
-import java.math.BigInteger;
-
 public interface Unterscheidbar {
 	String getName();
-	BigInteger getWert();
+	Integer getWert();
 }
